chore(tests): extract COMMENT_REPLY tests to their own file
Both test_in_app_notifications.py (521 lines) and test_event_comments.py
(501 lines) crossed the 500-line cap after adding the PR2 cases.

Pulls TestNotifyCommentReply and the reply-endpoint integration test
into a new test_comment_reply_notifications.py. The integration test
no longer depends on the rsvp_user / rsvp_headers / event_with_rsvp
fixtures from test_event_comments.py — it builds its own minimal
setup inline (one event, two RSVPd users) so it's self-contained.
