+++ backend/tests/test_in_app_notifications.py
@@ -3,14 +3,13 @@
 from unittest.mock import patch
 
 import pytest
-from community.models import Event, EventComment, PageVisibility
+from community.models import Event, PageVisibility
 from ninja_jwt.tokens import RefreshToken
 from notifications.models import Notification, NotificationType
 from notifications.service import (
     _notify_users,
     create_event_invite_notifications,
     create_join_request_notifications,
-    notify_comment_reply,
 )
 from users.models import User
 from users.permissions import PermissionKey
@@ -459,63 +458,3 @@ class TestPgNotifyIntegration:
         """_notify_users is a no-op when not on PostgreSQL (e.g. SQLite in tests)."""
         # Should not raise even though no pg_notify is available
         _notify_users(["fake-uuid"])
-
-
-# ─── Comment Reply Notification Tests ────────────────────────────────────────
-
-
-@pytest.mark.django_db
-class TestNotifyCommentReply:
-    def test_notifies_parent_author(self, test_user, db):
-        replier = User.objects.create_user(
-            phone_number="+12025550707",
-            password="replierpass",
-            display_name="Replier",
-        )
-        event = Event.objects.create(
-            title="E",
-            start_datetime=future_iso(days=10),
-            created_by=test_user,
-        )
-        parent = EventComment.objects.create(event=event, author=test_user, body="parent")
-        reply = EventComment.objects.create(
-            event=event, author=replier, body="reply", parent=parent
-        )
-        notify_comment_reply(reply)
-        notifs = Notification.objects.filter(
-            recipient=test_user, notification_type=NotificationType.COMMENT_REPLY
-        )
-        assert notifs.count() == 1
-        n = notifs.first()
-        assert n.event_id == event.id
-        assert n.related_user_id == replier.id
-        assert "replier" in n.message.lower() or "Replier" in n.message
-
-    def test_no_self_notify(self, test_user, db):
-        event = Event.objects.create(
-            title="E",
-            start_datetime=future_iso(days=10),
-            created_by=test_user,
-        )
-        parent = EventComment.objects.create(event=event, author=test_user, body="p")
-        reply = EventComment.objects.create(event=event, author=test_user, body="r", parent=parent)
-        notify_comment_reply(reply)
-        assert (
-            Notification.objects.filter(
-                recipient=test_user, notification_type=NotificationType.COMMENT_REPLY
-            ).count()
-            == 0
-        )
-
-    def test_noop_for_top_level(self, test_user, db):
-        event = Event.objects.create(
-            title="E",
-            start_datetime=future_iso(days=10),
-            created_by=test_user,
-        )
-        comment = EventComment.objects.create(event=event, author=test_user, body="top")
-        notify_comment_reply(comment)
-        assert (
-            Notification.objects.filter(notification_type=NotificationType.COMMENT_REPLY).count()
-            == 0
-        )
